requirements.txt
openai
requests
python-dotenv
pandas
numpy
nltk
loguru
langchain
langchain-community
#unstructured[all=docs]==0.12.0
pdf2image
langchain_openai
pinecone
pinecone-client
